retrieve parcel data with updated zoning types and add multiple zoning updates/deletes/creates
