+++ .env
@@ -1,4 +1,4 @@
-PRIMARY_DB_URL=jdbc:postgresql://localhost:5432/zoningtool
+PRIMARY_DB_URL=jdbc:postgresql://localhost:5432/zoningtooldb
 PRIMARY_DB_USERNAME=postgres
 PRIMARY_DB_PASSWORD=password
