+++ src/main/java/com/eliass/zoningtool/realestatedb/service/RealEstateZoningService.java
@@ -1,0 +1,16 @@
+package com.eliass.zoningtool.realestatedb.service;
+
+import com.eliass.zoningtool.realestatedb.dto.RealEstateZoningDto;
+import com.eliass.zoningtool.realestatedb.entity.RealEstateZoning;
+
+import java.util.List;
+
+public interface RealEstateZoningService {
+
+    RealEstateZoningDto getRealEstateZoningById(Integer zoningUpdateId);
+
+    List<RealEstateZoning> getAllRealEstateZoning();
+
+
+
+}
