--- src/main/java/com/eliass/zoningtool/realestatezoning/exception/ResourceNotFoundException.java
@@ -1,12 +1,0 @@
-package com.eliass.zoningtool.realestatezoning.exception;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(value = HttpStatus.NOT_FOUND)
-public class ResourceNotFoundException extends RuntimeException{
-
-    public ResourceNotFoundException(String message){
-        super(message);
-    }
-}
