--- src/main/java/com/eliass/zoningtool/realestatezoning/service/RealEstateZoningService.java
@@ -1,13 +1,0 @@
-package com.eliass.zoningtool.realestatezoning.service;
-
-import com.eliass.zoningtool.realestatezoning.dto.RealEstateZoningDto;
-
-import java.util.List;
-
-public interface RealEstateZoningService {
-
-    RealEstateZoningDto getRealEstateZoningById(Integer zoningUpdateId);
-
-    List<RealEstateZoningDto> getAllRealEstateZoning();
-
-}
