+++ src/main/java/com/eliass/zoningtool/zoningtooldb/exception/ResourceNotFoundException.java
@@ -1,0 +1,12 @@
+package com.eliass.zoningtool.zoningtooldb.exception;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(value = HttpStatus.NOT_FOUND)
+public class ResourceNotFoundException extends RuntimeException{
+
+    public ResourceNotFoundException(String message){
+        super(message);
+    }
+}
