--- src/main/java/com/eliass/zoningtool/zoningupdates/dto/ZoningUpdateDto.java
@@ -1,16 +1,0 @@
-package com.eliass.zoningtool.zoningupdates.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class ZoningUpdateDto {
-    private Integer id;
-    private Integer parcelId;
-    private String zoningTyp;
-}
