--- src/main/java/com/eliass/zoningtool/zoningupdates/repository/ZoningUpdateRepository.java
@@ -1,7 +1,0 @@
-package com.eliass.zoningtool.zoningupdates.repository;
-
-import com.eliass.zoningtool.zoningupdates.entity.ZoningUpdate;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ZoningUpdateRepository extends JpaRepository<ZoningUpdate, Integer> {
-}
