--- src/main/java/com/eliass/zoningtool/zoningupdates/service/ZoningUpdateService.java
@@ -1,17 +1,0 @@
-package com.eliass.zoningtool.zoningupdates.service;
-
-import com.eliass.zoningtool.zoningupdates.dto.ZoningUpdateDto;
-
-import java.util.List;
-
-public interface ZoningUpdateService {
-    ZoningUpdateDto createZoningUpdate(ZoningUpdateDto zoningUpdateDto);
-
-    ZoningUpdateDto getZoningUpdateById(Integer zoningUpdateId);
-
-    List<ZoningUpdateDto> getAllZoningUpdates();
-
-    ZoningUpdateDto updateZoningUpdate(Integer zoningUpdateId, ZoningUpdateDto updatedZoningUpdate);
-
-    void deleteZoningUpdate(Integer zoningUpdateId);
-}
